backend/tests/requirements.txt
# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.27.2
Pillow==10.4.0

# For integration tests
pytest-cov==4.1.0
pytest-xdist==3.5.0

# For performance testing
pytest-benchmark==4.0.0
